fix: remove the applyBabelUserconfig to @modern-js/utils
