+++ packages/builder/builder-rspack-provider/package.json
@@ -49,21 +49,16 @@
   },
   "devDependencies": {
     "@modern-js/e2e": "workspace:*",
-    "@modern-js/babel-preset-app": "workspace:*",
     "@scripts/vitest-config": "workspace:*",
     "@types/node": "^14",
     "postcss": "8.4.21",
     "typescript": "^4"
   },
   "peerDependencies": {
-    "@modern-js/e2e": "workspace:^2.9.0",
-    "@modern-js/babel-preset-app": "workspace:^2.9.0"
+    "@modern-js/e2e": "workspace:^2.9.0"
   },
   "peerDependenciesMeta": {
     "@modern-js/e2e": {
-      "optional": true
-    },
-    "@modern-js/babel-preset-app": {
       "optional": true
     }
   },
